# src/controllers/thread.controller.ts
import { NextFunction, Request, Response } from 'express';
import { v2 as cloudinary } from 'cloudinary';
import threadService from '../services/thread.service';
import { createThreadSchema } from '../utils/schemas/thread.schema';
import likesService from '../services/like.service';
import { uploadImageSingle } from '../utils/upload';

class ThreadController {
  // get all Thread
  async getThreads(req: Request, res: Response, next: NextFunction) {
    try {
      // login UserId
      const userId = req?.user?.id;

      const page = parseInt(req.query.page as string) || 1;
      const limit = parseInt(req.query.limit as string) || 10;
      const startIndex = (page - 1) * limit;

      const pagination = {
        page,
        limit,
        startIndex,
      };

      const threads = await threadService.getThreads(pagination);

      const newThreads = await Promise.all(
        threads.map(async (thread) => {
          const like = await likesService.getLike(userId, thread.id);
          const isLiked = like ? true : false;
          const likesCount = thread.likes.length;
          const repliesCount = thread.replies.length;

          return {
            ...thread,
            likesCount,
            repliesCount,
            isLiked,
          };
        })
      );

      res.json(newThreads);
    } catch (error) {
      next(error);
    }
    return;
  }

  // get one thread by Id
  async getThreadById(req: Request, res: Response, next: NextFunction) {
    try {
      const { id } = req.params;

      const thread = await threadService.getThreadById(id);

      if (!thread) {
        res.status(404).json({
          message: 'Thread is not found!',
        });
        return;
      }
      let isLiked = false;

      if (req.user) {
        const userId = req.user.id;
        const like = await likesService.getLike(userId, thread.id);
        isLiked = like ? true : false;
      }

      const likesCount = thread?.likes.length;
      const repliesCount = thread?.replies.length;

      res.json({
        ...thread,
        likesCount,
        repliesCount,
        isLiked,
      });
    } catch (error) {
      console.log('error here');
      console.log(error);
      next(error);
    }
    return;
  }

  async getThreadByUserId(req: Request, res: Response) {
    try {
      const { userId } = req.params;
      const thread = await threadService.getThreadByUserId(userId);
      res.json(thread);
    } catch (error) {
      res.json(error);
    }
    return;
  }

  // create thread
  async createThread(req: Request, res: Response, next: NextFunction) {
    /*  #swagger.requestBody = {
            required: true,
            content: {
              "multipart/form-data": {
                    schema: {
                        $ref: "#/components/schemas/CreateThreadDTO"
                    }  
                }
            }
        } 
    */
    try {
      let imageUrl: string = '';
      if (req.file) {
        imageUrl = await uploadImageSingle('dumbways/threads', req);
      }
      const body = {
        ...req.body,
        images: imageUrl || undefined,
      };

      const userId = req.user.id;
      const validatedBody = await createThreadSchema.validateAsync(body);
      const thread = await threadService.createThread(userId, validatedBody);

      console.log(thread);
      if (thread) {
        res.status(201).json({
          message: 'Thread created!',
          data: { ...thread },
        });
      } else {
        res.status(500).json({
          message: 'Thread failed to created!',
        });
      }
    } catch (error) {
      console.log('error coy');
      console.log(error);
      next(error);
    }
    return;
  }

  // update thread:
  async updateThread(req: Request, res: Response, next: NextFunction) {
    /*  #swagger.requestBody = {
            required: true,
            content: {
              "multipart/form-data": {
                    schema: {
                        $ref: "#/components/schemas/UpdateThreadDTO"
                    }  
                }
            }
        } 
    // */

    try {
      let body: object = {};
      let imageUrl: string = '';

      const data = await threadService.getThreadById(req.body.id);

      if (data?.images) {
        const images = data?.images
          .split('/')
          .splice(7, 4)
          .join('/')
          .split('.')[0];
        cloudinary.uploader.destroy(images, function (result) {
          console.log('Succesfuly deleted image!', result);
        });
      }

      if (req.file) {
        imageUrl = await uploadImageSingle('dumbways/threads', req);

        body = {
          ...req.body,
          images: imageUrl || undefined,
        };
      } else {
        body = req.body;
      }

      const userId = req.user.id;
      const validatedBody = await createThreadSchema.validateAsync(body);
      const thread = threadService.updateThread(userId, validatedBody);
      res
        .status(202)
        .json({ message: 'Succesfully updated the thread!', ...thread });
    } catch (error) {
      next(error);
    }
  }

  // delete thread
  async deleteThread(req: Request, res: Response, next: NextFunction) {
    try {
      const data = await threadService.getThreadById(req.body.id);

      if (data?.images) {
        const images = data?.images
          .split('/')
          .splice(7, 4)
          .join('/')
          .split('.')[0];
        cloudinary.uploader.destroy(images, function (result) {
          console.log('Succesfuly deleted image!', result);
        });
      }

      const thread = await threadService.deleteThread(req.body.id);
      res
        .status(204)
        .json({ message: 'Succesfully deleted the image!', ...thread });
    } catch (error) {
      next(error);
    }
  }
}

export default new ThreadController();